+++ automaton/IdAutomaton.h
@@ -9,7 +9,7 @@ using namespace std;
 class IdAutomaton : public Automaton {
 public:
     AutomatonResult s0(const string& input, size_t currIndex, int currLine) override;
-    AutomatonResult s1(const string& input, int initIndex, size_t currIndex, int currLine);
+    AutomatonResult s1(const string& input, size_t initIndex, size_t currIndex, int currLine);
     TestResult testAutomaton() override;
 };
